+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [8.2.1] - Unreleased
+
+- Add Release status options to the ticketing system
 - Improve Interface for Lookup table generation
 
 ## [8.2.0] - 2024-07-09